==> Notebooks/1.a)Data Processing/3).Feature Engineering.ipynb <==
import pandas as pd

df = pd.read_csv("All_Routes_Processed.csv")
pd.set_option('display.max_columns', None)

df.shape

## Creating Average Speed

# df['Average_Speed'] = df['distance_m'] / df['duration_s']

### Adding Turn Density

df['Turn_Density'] = df['turns'] / (df['distance_m'] / 1000)

### Surface Data

df["on_road"] = df["Asphalt"] + df["Concrete"]
df.drop(columns=["Asphalt", "Concrete"], inplace=True)

df["off_road"] = df["Dirt"] + df["Grass"] + df["Sand"] + df["Ground"] +df["Unpaved"]
df.drop(columns=["Dirt", "Grass", "Sand", "Ground", "Unpaved"], inplace=True)

df["Gravel_Tracks"] = df["Gravel"] + df["Compacted Gravel"]
df.drop(columns=["Gravel", "Compacted Gravel"], inplace=True)

df["Paved_Paths"] = df["Paved"] + df["Paving Stones"]
df.drop(columns=["Paved", "Paving Stones"], inplace=True)

df["Other"] = df["Wood"] + df["Metal"] + df["Grass Paver"]
df.drop(columns=["Wood", "Metal", "Grass Paver"], inplace=True)

df["Unknown Surface"] = df["Unknown"]
df.drop(columns=["Unknown"], inplace=True)

### Way Types

df["Paved_Road"] = df["Road"] + df["Street"]
df.drop(columns=["Road", "Street"], inplace=True)

df["Pedestrian"] = df["Steps"] + df["Footway"]
df.drop(columns=["Steps", "Footway"], inplace=True)

df["Unknown_Way"] = df["Unknown.1"]
df.drop(columns=["Unknown.1"], inplace=True)

df["Cycle Track"] = df["Path"] + df["Track"]
df.drop(columns=["Path", "Track"], inplace=True)

df["Other"] = df["Ferry"] + df["Construction"]
df.drop(columns=["Ferry", "Construction"], inplace=True)

df["Main Road"] = df["State Road"]
df.drop(columns=["State Road"], inplace=True)

### Steepness

df["Steep Section"] = df["uphill_steep (5% to 7%)"] + df["uphill_very_steep (7% to 10%)"] + df["uphill_extreme (>10%)"]
df.drop(columns=["uphill_steep (5% to 7%)", "uphill_very_steep (7% to 10%)", "uphill_extreme (>10%)"], inplace=True)

df["Moderate Section"] = df["uphill_gentle (0% to 3%)"] + df["uphill_moderate (3% to 5%)"]
df.drop(columns=["uphill_gentle (0% to 3%)", "uphill_moderate (3% to 5%)"], inplace=True)

df["Flat Section"] = df["flat (0%)"]
df.drop(columns=["flat (0%)"], inplace=True)

df["Downhill Section"] = df["downhill_gentle (-5% to 0%)"] + df["downhill_moderate (-7% to -5%)"]
df.drop(columns=["downhill_gentle (-5% to 0%)", "downhill_moderate (-7% to -5%)"], inplace=True)

df["Steep Downhill Section"] = df["downhill_steep (-10% to -7%)"] + df["downhill_very_steep (-15% to -10%)"] + df["downhill_extreme (<-15%)"]
df.drop(columns=["downhill_steep (-10% to -7%)", "downhill_very_steep (-15% to -10%)", "downhill_extreme (<-15%)"], inplace=True)

### Dataset

df

df.to_csv('Data_Engineered.csv', index=False)

df
